backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
PyYAML==6.0.1